requirements.txt
# RunPod Serverless Worker Requirements
runpod>=1.6.0
torch>=2.0.0
torchaudio>=2.0.0
umap-learn>=0.5.0
numpy>=1.24.0
scipy>=1.10.0
numba>=0.57.0
# stable-audio-tools installed from git in Dockerfile
